+++ src/Services/RemoteEventsFabric.php
@@ -13,14 +13,16 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Services;
 
+use Bitrix24\SDK\Application\Requests\Events\ApplicationLifeCycleEventsFabric;
 use Bitrix24\SDK\Core\Contracts\Events\EventInterface;
 use Bitrix24\SDK\Core\Contracts\Events\EventsFabricInterface;
 use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
-use Bitrix24\SDK\Core\Requests\Events\UnsupportedEvent;
+use Bitrix24\SDK\Core\Requests\Events\UnsupportedRemoteEvent;
+use Bitrix24\SDK\Services\Telephony\Events\TelephonyEventsFabric;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\Request;
 
-readonly class EventsFabric
+readonly class RemoteEventsFabric
 {
     /**
      * @throws InvalidArgumentException
@@ -41,6 +43,22 @@ readonly class EventsFabric
     }
 
     /**
+     * Check is remote events fabric can process this request and create event object
+     *
+     * @param Request $request
+     * @return bool
+     */
+    public function isCanProcess(Request $request): bool
+    {
+        $payload = [];
+        parse_str($request->getContent(), $payload);
+
+        return array_key_exists('event', $payload);
+    }
+
+    /**
+     * Create events objects from incoming events from Bitrix24
+     *
      * @throws InvalidArgumentException
      */
     public function createEvent(Request $request): EventInterface
@@ -55,7 +73,7 @@ readonly class EventsFabric
             'eventCode' => $eventCode
         ]);
 
-        $event = new UnsupportedEvent($request);
+        $event = new UnsupportedRemoteEvent($request);
         foreach ($this->eventsFabrics as $itemFabric) {
             if ($itemFabric->isSupport($eventCode)) {
                 $event = $itemFabric->create($request);
@@ -68,5 +86,26 @@ readonly class EventsFabric
             'eventCode' => $event->getEventCode()
         ]);
         return $event;
+    }
+
+    /**
+     * Initializes the EventsFabric with the given logger.
+     *
+     * @param LoggerInterface $logger The logger to be used for logging events.
+     *
+     * @return self The initialized EventsFabric instance.
+     *
+     * @throws InvalidArgumentException If any of the events fabrics in the array do not implement the EventsFabricInterface.
+     */
+    public static function init(LoggerInterface $logger): self
+    {
+        return new self(
+            [
+                // register events fabric by scope
+                new ApplicationLifeCycleEventsFabric(),
+                new TelephonyEventsFabric(),
+            ],
+            $logger
+        );
     }
 }