+++ tests/Unit/Services/CRM/CRMServiceBuilderTest.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace Bitrix24\SDK\Tests\Unit\Services\CRM;
 
 use Bitrix24\SDK\Services\CRM\CRMServiceBuilder;
-use Bitrix24\SDK\Services\EventsFabric;
+use Bitrix24\SDK\Services\RemoteEventsFabric;
 use Bitrix24\SDK\Services\ServiceBuilder;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullBatch;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullBulkItemsReader;
@@ -69,7 +69,6 @@ class CRMServiceBuilderTest extends TestCase
             new NullCore(),
             new NullBatch(),
             new NullBulkItemsReader(),
-            new EventsFabric([], new NullLogger()),
             new NullLogger()
         ))->getCRMScope();
     }