+++ tests/Unit/Services/IM/IMServiceBuilderTest.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Tests\Unit\Services\IM;
 
-use Bitrix24\SDK\Services\EventsFabric;
+use Bitrix24\SDK\Services\RemoteEventsFabric;
 use Bitrix24\SDK\Services\IM\IMServiceBuilder;
 use Bitrix24\SDK\Services\ServiceBuilder;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullBatch;
@@ -40,7 +40,6 @@ class IMServiceBuilderTest extends TestCase
             new NullCore(),
             new NullBatch(),
             new NullBulkItemsReader(),
-            new EventsFabric([],new NullLogger()),
             new NullLogger()
         ))->getIMScope();
     }